requirements.txt
Flask==3.1.2
google-genai==1.31.0
pillow==11.3.0
python-dotenv==1.0.0